Refactors upload batch and response handling
Extracts batch-specific operations, including collection creation, retry logic, and recovery mechanisms, into a new `useUploadBatchOperations` composable. This improves modularity and separates concerns from the main `useUploadQueue`.
